Fixed issue with duplicate rows (culprit: IMF)
There were some duplicate rows for 2023 in quite a few datasets, which was due to duplicates in the IMF data, which were introduced via the custom added column "ObservationType", which is now removed and the duplicates have also disappeared as a result. I've refresh all data. I don't expect any changes in the results, perhaps the predictions, but the rest should be as it currently is.